Move build_read_books_all_users_dict() from EvaluationDataProvider to EvaluationPopularity
