+++ RecommenderModule/evaluation/resources/evaluation_data_provider.py
@@ -7,7 +7,6 @@ class EvaluationDataProvider:
     dataset = None
     loocv_trainset = None
     loocv_testset = None
-    read_books_all_users = {}
     
     def __init__(self, dataset):
         self.dataset = dataset
@@ -27,24 +26,3 @@ class EvaluationDataProvider:
         the original dataset, using 'Leave-One-Out'."""
     def get_loocv_datasets(self):
         return (self.loocv_trainset, self.loocv_testset)
-    
-    
-    """Build the read_books_all_users dictionary, which for every user holds 
-        an array of all books read by that user."""
-    def build_read_books_all_users_dict(self):
-        self.read_books_all_users = {}
-        trainset = self.loocv_trainset
-        for inner_user_id, inner_item_id, rating in trainset.all_ratings():
-            raw_user_id = trainset.to_raw_uid(inner_user_id)
-            raw_item_id = trainset.to_raw_iid(inner_item_id)
-            try:
-                self.read_books_all_users[raw_user_id].append(raw_item_id)
-            except:
-                self.read_books_all_users[raw_user_id] = [raw_item_id]
-    
-    """Returns the read_books_all_users dictionary containing the books read
-        by every user."""
-    def get_read_books_all_users_dict(self):
-        if (self.read_books_all_users == {}):
-            self.build_read_books_all_users_dict()
-        return self.read_books_all_users
